Move header and search bar into folders to organise components directory
