+++ src/pages/CocktailListPage.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import SearchBar from '../components/SearchBar';
+import SearchBar from '../components/SearchBar/SearchBar';
 import CocktailList from '../components/CocktailList/CocktailList';
 
 const CocktailListPage = () => {
